fix: Enable full GPU for SNARKer (#60)
SNARK prover has 3 main steps:
- merge fris
- final proof
- snark wrapping

Of those, on GPU flag we had a mix of GPU, CPU and GPU again. This PR
makes it all GPU (GPU, GPU, GPU). This setup requires 32 GB of VRAM, but
will be decreased in the future (see the note in comments).

This change will allow deployed nets to run ~3 minutes faster. If you
need to run locally, you may change the flags to adjust for your own
hardware.

NOTE: Previous changes removed default GPU flag for SNARK as it was
applying it by default for FRI. This commit contains a change that also
builds the SNARK with gpu features (as it's not default anymore).