docs: Add documentation & setup for ubuntu VM
Add `docs/` folder with `setup_linux_vm.md` to explain the process of
getting the entire suite of testing working. Additionally, `scripts/`
folder is added that contains `ubuntu_setup.sh` to speedrun through some
of the setup process.

This is the boilerplate, we can expand on this further in the future.