feat: Improve pick SNARK UX
Previously, the line would state that the job was saved and then just
hang. In fact it was saving after it stated the saving was done.
Eventually it would finish. This change states when the job is picked
and that it will save afterwards. And once save to disk is written
notifies.

Furthermore, adds sensible sequencer url for local development.