feat: Add ZK to SNARK
The PR updates the entire proving stack:
- zkos-wrapper = v0.5.2
- zksync-crypto-gpu = 0.155.2
- zksync-protocol = 0.153.3
- zksync-crypto = 0.32.5

Airbender is untouched. New changes were added to crypto to allow
padding proofs that have been propragated through out the stack. This
allows real ZK to our SNARKs. On top of it, there are a few other minor
changes that were not done (such as passing references instead of
cloning -- saves us a bunch of allocation).

This is needed as part of Prividium release.