feat: Protocol Upgrade support (#96)
PR adds support for protocol upgrades and further troubleshooting.

This is v0.0.1 of the implementation. It allows provers to only jobs for
which they are designed. It also documents current protocol version
usage. On follow-up PR, we can add multiple protocol versions to be
supported at a time (i.e. when changes are applied only to zksync-os).

NOTE: These changes are NOT backwards compatible, as that would require
more complexity on the prover side that would only be needed after
deployment. The backwards compatible part is done on server side. As
such, the expectation is that the server will deployed, followed by
prover.
NOTE2: The changes here correspond to interface changes in server
[here](https://github.com/matter-labs/zksync-os-server/pull/577).
NOTE3: Has been tested locally, seems to be working properly.